requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
config==0.5.1
contourpy==1.3.1
contrib==0.3.0
cycler==0.12.1
fire==0.7.0
fonttools==4.56.0
fpdf==1.7.2
frozenlist==1.5.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.1
magic-filter==1.0.12
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pdf2docx==0.5.8
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
setuptools==75.8.0
six==1.17.0
termcolor==2.5.0
typing_extensions==4.12.2
yarl==1.18.3
